+++ internal/flags/options/options.go
@@ -19,6 +19,8 @@ type Options struct {
 	DefaultMaxComputeMinutesPerWeek                 float64
 	DefaultApiTimeout                               float64
 	DefaultGoWorkerEnabled                          bool
+	DefaultEphemeralEnabledPlugins                  []string
+	DefaultEphemeralSupportedEvents                 []string
 	DefaultWorkflowPluginInheritanceEnabled         bool
 	DefaultValidateSubjectTokenDuringOboFlowEnabled bool
 	Logger                                          *zap.Logger
@@ -99,6 +101,18 @@ func WithDefaultGoWorkerEnabled(goWorkerEnabled bool) Option {
 	}
 }
 
+func WithDefaultEphemeralEnabledPlugins(ephemeralEnabledPlugins ...string) Option {
+	return func(d *Options) {
+		d.DefaultEphemeralEnabledPlugins = ephemeralEnabledPlugins
+	}
+}
+
+func WithDefaultEphemeralSupportedEvents(ephemeralSupportedEvents ...string) Option {
+	return func(d *Options) {
+		d.DefaultEphemeralSupportedEvents = ephemeralSupportedEvents
+	}
+}
+
 func WithDefaultWorkflowPluginInheritanceEnabled(workflowPluginInheritanceEnabled bool) Option {
 	return func(d *Options) {
 		d.DefaultWorkflowPluginInheritanceEnabled = workflowPluginInheritanceEnabled
@@ -130,6 +144,8 @@ func Apply(opts ...Option) Options {
 		DefaultMaxComputeMinutesPerWeek:         float64(math.MaxInt32),
 		DefaultApiTimeout:                       float64(time.Duration(1 * time.Hour).Milliseconds()), // NOTE(frank): This is arbitrary.
 		DefaultGoWorkerEnabled:                  false,
+		DefaultEphemeralEnabledPlugins:          []string{},
+		DefaultEphemeralSupportedEvents:         []string{},
 		DefaultWorkflowPluginInheritanceEnabled: false,
 		Logger:                                  zap.NewNop(),
 		Config:                                  nil,