+++ pkg/worker/options/options.go
@@ -4,6 +4,8 @@ type Options struct {
 	Stream                    chan<- string
 	ApiTimeoutErrorPrecedence bool
 	ApiTimeoutError           error
+	OrganizationPlan          string
+	OrgId                     string
 }
 
 type Option func(*Options)
@@ -28,5 +30,17 @@ func ApiTimeoutErrorPrecedence(b bool, err error) Option {
 	return func(o *Options) {
 		o.ApiTimeoutErrorPrecedence = b
 		o.ApiTimeoutError = err
+	}
+}
+
+func OrganizationPlan(plan string) Option {
+	return func(o *Options) {
+		o.OrganizationPlan = plan
+	}
+}
+
+func OrgId(id string) Option {
+	return func(o *Options) {
+		o.OrgId = id
 	}
 }