+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Restrict access to blocked packages in JavaScript steps via: dynamic import, `eval`, and Function constructors
 - Update JavaScript worker to execute language steps within `vm2`
 - Add support for Kubernetes Gateway API as an alternative to Ingress for routing traffic to the agent
+- Update FilePicker download endpoint to ensure requested download paths are under the temp directory
 
 ## v1.31.0
 - Add support for restricting imports and built in functions in Python language execution steps